backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.1.0
pydantic>=2.7.0
pydantic-settings[toml]>=2.2.0
pydantic-ai>=0.0.9
mcp[cli]>=1.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.30
alembic>=1.13.0
pgvector>=0.3.0
cryptography>=42.0.0
openai>=1.30.0
anthropic>=0.25.0
google-genai>=1.0.0
celery[redis]>=5.4.0
flower>=2.0.0
redis>=5.0.0
httpx>=0.27.0
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-typescript>=0.21.0
mistune>=3.0.0
gitpython>=3.1.40
python-multipart>=0.0.9
python-dateutil>=2.9.0
tiktoken>=0.7.0
croniter>=2.0.0
jmespath>=1.0.0
boto3>=1.34.0
google-cloud-storage>=2.16.0
